refactor(ChessGameDBService): 이름 변경 및 ChessGameDBService에서 도메인을 반환하도록 변경
