+++ docs/README.md
@@ -83,8 +83,8 @@ abcdefgh
 - [ ] README.md 어플리케이션 실행 방법 정리
 - [x] ScoreManager ↔️ ChessState 객체끼리 서로 물어보며 점수 계산
 - [ ] ScoreManager 상수 사용 or 각 객체에서 자신의 점수 계산
-- [ ] DBService 네이밍
-  - [ ] 무엇을 위한 서비스인지 고민
-  - [ ] 도메인을 서비스에서 만들어서 반환
+- [x] DBService 네이밍
+  - [x] 무엇을 위한 서비스인지 고민
+  - [x] 도메인을 서비스에서 만들어서 반환
 - [ ] DBConnector 네이밍 & 역할
 - [x] PieceType : 생성자 사용으로 통일
